requirements.txt
pylint==3.3.1
pytest==8.3.3
python-dotenv==1.0.1
Flask==3.0.3
SQLAlchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1
PyJWT==2.9.0
pyroute2==0.7.12
wgconfig==1.0.4
mongomock==4.2.0.post1
mongoengine==0.29.1
docker==7.1.0
